Build: Publish theme as package
By publishing Amethyst as a package, and using it by name in individual
website repos (instead of by its raw Git repo URL), the dependencies
will be managed automatically, which should make it easier to get
started, and more importantly, easier to update.

(The initial start is equally easy as it involved copying the example
boilerplate, but the steps after that are less obvious).